+++ index.html
@@ -233,7 +233,6 @@
   </template>
 
   <!-- Подключение скриптов -->
-  <script src="/js/function.js"></script>
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
 </body>
 </html>
