+++ js/constants.js
@@ -1,0 +1,54 @@
+export const POSTS_COUNT = 25;
+export const LIKES_COUNT = { MIN: 15, MAX: 200 };
+export const AVATAR_COUNT = 6;
+export const COMMENTS_COUNT = 30;
+
+export const POST_DESCRIPTIONS = [
+  'Закат на пляже',
+  'Городская панорама ночью',
+  'Зимний лес',
+  'Макро съемка цветка',
+  'Архитектурные детали здания',
+  'Пейзаж с горами',
+  'Городской фонарь в тумане',
+  'Мост через реку',
+  'Парк с вишневыми деревьями весной',
+  'Портрет девушки в черно-белом стиле',
+  'Собака играет на пляже',
+  'Закат в горах',
+  'Абстрактное искусство',
+  'Расцветающее поле под солнцем',
+  'Старая улица в Европе',
+  'Макро съемка бабочки',
+  'Природа весной',
+  'Пейзаж с водопадом',
+  'Городская архитектура',
+  'Портрет мужчины',
+  'Зимний город ночью',
+  'Рыбак на рассвете',
+  'Пейзаж с озером',
+  'Рабочий процесс в офисе',
+  'Автомобиль на закате',
+];
+
+export const NAMES = [
+  'Анна',
+  'Иван',
+  'Мария',
+  'Александр',
+  'Екатерина',
+  'Дмитрий',
+  'Ольга',
+  'Михаил',
+  'Елена',
+  'Павел'
+];
+
+export const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
